refactor(tempo): merge channel descriptor
