+++ .changeset/loud-keys-float.md
@@ -1,0 +1,5 @@
+---
+"ox": patch
+---
+
+Renamed `ChannelDescriptor.from` to `Channel.from`, made `Channel.Channel` the descriptor type, and changed `Channel.computeId` to receive channel and options separately.
